server/requirements.txt
flask>=3.0.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
flask-cors>=4.0.0
